Users에서 결과값을 세팅하도록 변경
