+++ src/main/java/nextstep/ladder/LadderController.java
@@ -2,7 +2,6 @@ package nextstep.ladder;
 
 import nextstep.ladder.domain.GameInfo;
 import nextstep.ladder.domain.Ladder;
-import nextstep.ladder.domain.PlayLadderGame;
 import nextstep.ladder.domain.Results;
 import nextstep.ladder.domain.Users;
 import nextstep.ladder.view.InputView;
@@ -11,7 +10,6 @@ import nextstep.ladder.view.ResultView;
 public class LadderController {
     private static InputView inputView = InputView.getInputView();
     private static ResultView resultView = ResultView.getResultView();
-    private static PlayLadderGame playLadderGame = PlayLadderGame.getPlayLadderGame();
 
     public static void main(String[] args) {
         Users users = inputView.enterUserNames();
@@ -21,7 +19,7 @@ public class LadderController {
 
         Ladder ladder = new Ladder(users.getCountOfPerson(), ladderHeight);
         resultView.printLadder(gameInfo, ladder);
-        Users resultsForAllPlayers = playLadderGame.generateResultsForAllPlayers(gameInfo, ladder);
-        resultView.repeatPrintPlayResult(resultsForAllPlayers);
+        users.generateResultsForAllPlayers(gameInfo, ladder);
+        resultView.repeatPrintPlayResult(users);
     }
 }
