+++ src/main/java/nextstep/ladder/domain/Users.java
@@ -1,5 +1,6 @@
 package nextstep.ladder.domain;
 
+import java.util.Collections;
 import java.util.List;
 
 import nextstep.ladder.domain.User;
@@ -8,11 +9,29 @@ public class Users {
     private List<User> users;
 
     public Users(List<User> users) {
-        this.users = users;
+        this.users = Collections.unmodifiableList(users);
     }
 
     public int getCountOfPerson() {
         return users.size();
+    }
+
+    public void generateResultsForAllPlayers(GameInfo paramGameInfo, Ladder paramLadder) {
+        List<LadderLine> ladder = paramLadder.getLadder();
+        List<Result> results = paramGameInfo.getResults().getResults();
+
+        for (int i = 0; i < getCountOfPerson(); i++) {
+            int index = i;
+            index = getIndex(ladder, index);
+            users.get(i).setResult(results.get(index).getResult());
+        }
+    }
+
+    private int getIndex(List<LadderLine> ladder, int index) {
+        for (int j = 0; j < ladder.size(); j++) {
+            index = ladder.get(j).move(index);
+        }
+        return index;
     }
 
     public List<User> getUsers() {
